+++ src/portal.ts
@@ -54,10 +54,10 @@ export class PortalDirective extends AsyncDirective {
    * This directive always returns `nothing` because nothing ever renders where the portal is used.
    */
   render(
-    value: unknown | Promise<unknown>,
+    content: unknown | Promise<unknown>,
     targetOrSelector: TargetOrSelector | Promise<TargetOrSelector>,
   ) {
-    Promise.all([value, targetOrSelector]).then(([value, targetOrSelector]) => {
+    Promise.all([content, targetOrSelector]).then(([content, targetOrSelector]) => {
       if (!targetOrSelector) {
         throw Error(
           "Target was falsy. Are you using a Lit ref before its value is defined? If so, try using Lit's @queryAsync decorator instead (https://lit.dev/docs/api/decorators/#queryAsync).",
@@ -73,7 +73,7 @@ export class PortalDirective extends AsyncDirective {
         this.target = target;
       }
 
-      litRender(value, this.container);
+      litRender(content, this.container);
     });
 
     return nothing;